fix: Tile: drawTileString moved to tile-graphics TU
- drawTileString() and tileColor() moved to tile-graphics TU.
