+++ src/headers/tile-graphics.hpp
@@ -1,4 +1,11 @@
 #ifndef TILEGRAPHICS_H
 #define TILEGRAPHICS_H
 
+#include <string>
+
+namespace Game {
+struct tile_t;
+std::string drawTileString(tile_t currentTile);
+} // namespace Game
+
 #endif
