+++ src/headers/tile.hpp
@@ -4,14 +4,10 @@
 #include "global.hpp"
 
 namespace Game {
-
 struct tile_t {
   ull value{};
   bool blocked{};
 };
-
 } // namespace Game
-
-std::ostream &operator<<(std::ostream &os, const Game::tile_t &t);
 
 #endif
